Add sys package
This commit adds the 'sys' package including interfaces
for a shell runner, mounter, logger and syscalls. Most of the
code is based on parts of elemental-toolkit/v2 'types' package.

System interface implementations are split into individual
subpackages.

Finally it also includes a Makefile target to run the unit tests and
compute coverage.

Signed-off-by: David Cassany <dcassany@suse.com>