Extend runner interface and add rsync package
This commit extends Runner interface with a couple of
additional methods to support running commands with an
associated Context. It also provides a new method to parse
in real-time stdout and stderr.

The rsync package is wrapper for rsync command. It can
be initiated with a Context in case the caller considers
some sort of cancellation logic.

Signed-off-by: David Cassany <dcassany@suse.com>