Keep the FS interface and vfs together
This commit moves all FS logic and interface definition
to vfs package. This keeps package purpose easier to read
and prevents having a sys package with vague purpose.

Sys package is just a collection of certain system utilities
and the fact they are interfaces is just an encapsulation
to allow easy testing and to enforce a proper encapsulation
in code.

We are not aiming to build a sys library that can be imported
without forcing to also import specific implementations of the
interfaces it uses.

Signed-off-by: David Cassany <dcassany@suse.com>