Fix pipes management in Runner
This commit fixes pipes management in Runner to
ensure pipes are closed after all the data is read.

There was a race condition in pipes closing and,
moreover, a pipe can only be closed once, closing an
alredy closed pipe raises an error. Former code
was likely to close pipes twice.

Signed-off-by: David Cassany <dcassany@suse.com>