Force formatting even if a filesystem already exists
Forcing the format is required in some corner cases,
for instance if we run the same installation over the
same device twice. In that case the repartitioning
does not delete the previous filesystem as there is
a perfect match and then mkfs tools complains of
trying to format an already existing filesystem.

Signed-off-by: David Cassany <dcassany@suse.com>