Add unit-tests for local OCI unpacker
This commit adds unit-tests for unpacking OCI images using the OCI
unpacker with the local flag.

We test that:

1. pulling and unpacking a local docker image works
2. unpacking a local bogus image returns error
3. unpacking a remote bogus image returns error

This also introduces a 'docker pull ...' side-effect for test 1, which
depends on docker being present on the machine and able to pull
images from docker.io (same as old remote tests).

Signed-off-by: Fredrik Lönnegren <fredrik.lonnegren@suse.com>