Omit device name on deployment marshalling
This commit omits the disk device name when marshalling
to a json or yaml. The reason for that is to avoid storing
the device name into the deployment reference file.
The device names (e.g. /dev/sda ) might not be consistent
across reboots, hence we can't assume this is an static value
after the installation.

Signed-off-by: David Cassany <dcassany@suse.com>