Remove sector size from deployment struct
This commit removes sector size from the deployment struct
because this is not a deployment parameter, this is a given
from the context and hardware.

In installation scope sector size can't be customized, it is
a parameter of the target device. Sector size is only potentially
configurable in a disk image building context.

Signed-off-by: David Cassany <dcassany@suse.com>