Add start sector field in deployment
This commit adds a startSector attribute in Disk
struct. This is used to set the first sector in disk
that is used for a partition. This is specially relevant
for some ARM devices requiring the ESP partition at a
very specific position.

If unset defaults to 1MiB which is 2048 for disks with
a sector size of 512 bytes.

Signed-off-by: David Cassany <dcassany@suse.com>