Set partition alignement
Sgdisk uses 1MiB alignement by default on partition start
but not on partition end. Partitions partitions are always defined
as a multiple of 1MiB with the only exception of the last partition,
which takes all available space.

This commit ensures the last partition is also aligned to 1MiB. This
turns to be a requirement partition encryption tools.

Signed-off-by: David Cassany <dcassany@suse.com>