Add new boot entries on upgrade
The current 'active' entry will point to the latest snapshot and during
upgrade new boot entries are appended, leading to a log of old snapshots
being kept.

Also embed the grub config in bootloader package instead of extracting
it from image. The config is specific to the elemental toolkit and
should be packaged together for easier development.

Signed-off-by: Fredrik Lönnegren <fredrik.lonnegren@suse.com>