Drop block device detection based on ghw
This commit drops the implementation of the block interface
based on jaypipes/ghw library. This specific implementation is not
used anywhere in the code, it was just ported from classic
Elemental and meant to be the quick goto alternative of lsblk
in case of facing issues.

The reason to prefer the lsblk wrapper are diverse, but most prominent
is that lsblk is powerful and it is designed to be parsed. It relies
on udev (part of util-linux-systemd) and it is a well known tool
in the Linux community.

Dropping ghw based implementation is also reducing the libraries
elemental-toolkit depends on. Moreover some of the functionalities
of ghw also depend on udev database, hence using it does not really
reduce the host dependencies. By dropping it we have less code to
maintain.

Signed-off-by: David Cassany <dcassany@suse.com>