Add --local flag to install, upgrade and build commands
This flag forces the OCI unpack process to look for the given
image in the local container storage. The library expects a
daemon accessible through a socket typically defined by the
DOCKER_HOST environment variable.

This commit also changes the behavior of attempting to find
the image first in the local storage and fallback to remote if
not present. With this change there is no fallback logic and
pulling a remote or local image is an explicit choice.

Signed-off-by: David Cassany <dcassany@suse.com>