Add quick-start section to README
The quick-start aims to easily build and boot a freely available OS
(openSUSE Tumbleweed) with as few moving parts as possible.

Right now the `build`-command is not flexible enough to generate an
image with only freely available components, so we have to use the
losetup trick to create the image.

Signed-off-by: Fredrik Lönnegren <fredrik.lonnegren@suse.com>