Adding optional recovery partition management
This commit introduces the notion of a recovery partition.
It essentially sets an ISO installer to include an additional
partition after the ESP partition to include the same contents
the ISO filesystem has. This partition is bootable as a live
system and it sets the base to eventually be used in a potential
reset process to restore the system as it was installed back in
a day.

As part of the changes it includes a bootloader interface
refactor to drop its deployment package dependency. This to make
it easier to reuse it in different contexts, such as installer builds,
upgrades and clean installations.

Signed-off-by: David Cassany <dcassany@suse.com>