Add bridge networking to scripts/run_vm.sh
If invoked with ELMNTL_BRIDGE=<interface name> run_vm.sh will not try to
connect the spawned VM to the bridge, instead of doing the userspace
net.

We also add ELMNTL_MAC that can be used to set the MAC of the NIC.

This simplifies testing of the different node-types and allows for VMs
using DHCP over the bridge.

Signed-off-by: Fredrik Lönnegren <fredrik.lonnegren@suse.com>