K8s init node
InitServerConfig will be used by the initiating server to kick-start the
cluster deployment.

This also adds a type-alias ConfigMap for the 'map[string]any' used in
cluster.go for all the configuration maps.

The init node is the first server node that other nodes will join. It
has an empty 'server' config.

Signed-off-by: Fredrik Lönnegren <fredrik.lonnegren@suse.com>