Install rke2-selinux policy during firstboot
RKE2 is unable to start properly if it is enabled with the `selinux: true` configuration option. This is due to the systemd extension, containing both RKE2 itself and rke2-selinux policy, coming with properly relabeled files, resulting in improper context for the containerd and containerd-shim processes. By manually injecting the selinux policy this is now fixed.

Signed-off-by: Atanas Dinov <atanas.dinov@suse.com>