Do not store disk devices in recovery's install.yaml
This commit prevents to store the disk devices in
install.yaml when building or installing disks.

The reason for that is because when booting from a recovery
partition the disk devices might not match what they were at
install or build time. To prevent any undesired installation
or manipulation of a wrong device they are not included.

Signed-off-by: David Cassany <dcassany@suse.com>